+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 This project follows semantic versioning.
 
 ### Unpublished
-    - [changed] Removed dependency on libstd. ([Issue #53](https://github.com/paholg/typenum/issues/53), [PR #55](https://github.com/paholg/typenum/pull/55))
+    - [changed] Removed dependency on libstd. (Issue #53, PR #55)
+    - [changed] Reorganized module structure. (PR #56)
 
 
 ### 1.2.0 (2016-01-03)
     - [added] This change log!
-    - [added] Convenience type aliases for operators. ([Issue #48](https://github.com/paholg/typenum/issues/48), [PR #50](https://github.com/paholg/typenum/pull/50))
-    - [added] Types in this crate now derive all possible traits. ([Issue #42](https://github.com/paholg/typenum/issues/48), [PR #51](https://github.com/paholg/typenum/pull/51))
+    - [added] Convenience type aliases for operators. (Issue #48, PR #50)
+    - [added] Types in this crate now derive all possible traits. (Issue #42, PR #51)